+++ src/Session.cpp
@@ -69,7 +69,7 @@ void Session::Send( uptr<Buffer> pBuffer )
     write_token->writer         = new uv_write_t( );
     write_token->buffer         = new uv_buf_t( );
     write_token->buffer->base   = new char[data->Size( )] { 0 };
-    write_token->buffer->len    = data->Size( );
+    write_token->buffer->len    = (unsigned long)data->Size( );
     write_token->writer->data   = write_token;
     write_token->session        = this;
